% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcount.R
\name{calcount}
\alias{calcount}
\title{Calculate the number of calories of a food item based on protein, carb, and fat data}
\usage{
calcount(name, amount, unit)
}
\arguments{
\item{name}{Name of the food item}

\item{amount}{Number of servings}

\item{unit}{Unit of measure}
}
\value{
The number of calories consumed given the food type as well as protein, carb, and fat information
}
\description{
Calculate the number of calories of a food item based on protein, carb, and fat data
}

