% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcount.R, R/calcount2.R
\name{calcount2}
\alias{calcount2}
\title{Calculate the number of calories of a food item based on protein, carb, and fat data}
\usage{
calcount2(name, amount, unit, file = "R Package Data - Food1.csv")

calcount2(name, amount, unit, file = "R Package Data - Food1.csv")
}
\arguments{
\item{name}{Name of the food item}

\item{amount}{Number of servings}

\item{unit}{Unit of measure}

\item{name}{Name of the food item}

\item{amount}{Number of servings}

\item{unit}{Unit of measure}
}
\value{
The number of calories consumed given the food type as well as protein, carb, and fat information

The number of calories consumed given the food type as well as protein, carb, and fat information
}
\description{
Calculate the number of calories of a food item based on protein, carb, and fat data

Calculate the number of calories of a food item based on protein, carb, and fat data
}

