% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrition.R
\name{get_calories}
\alias{get_calories}
\title{Gets the calories for a recipe using the Nutritionix API}
\usage{
get_calories(text)
}
\description{
This function gets the calories for a recipe using the Nutritionix API
}
