% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{get_recipe_detail}
\alias{get_recipe_detail}
\title{Gets the details for a single recipe from Paprika using the recipe's UID}
\usage{
get_recipe_detail(uid)
}
\description{
This function will get the details for a Paprika recipe using UID.
}
