% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{get_recipes_list}
\alias{get_recipes_list}
\title{Gets a list of recipes from your Paprika account}
\usage{
get_recipes_list()
}
\description{
This function will fetch a list of recipes from Paprika account
}
