% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_creds.R
\name{install_nutritionix_creds}
\alias{install_nutritionix_creds}
\title{Adds Nutritionix App ID and API key to .Renviron file}
\usage{
install_nutritionix_creds(app_id, api_key)
}
\description{
This function will add your Nutritionix App ID and API key to your \code{.Renviron} file. Code for this function was used from a similar function in `tidycensus` (https://github.com/walkerke/tidycensus).
}
