% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_creds.R
\name{install_paprika_creds}
\alias{install_paprika_creds}
\title{Adds Paprika username and password to .Renviron file}
\usage{
install_paprika_creds(username, password)
}
\description{
This function will add your Paprika username and password to your \code{.Renviron} file. Code for this function was used from a similar function in `tidycensus` (https://github.com/walkerke/tidycensus).
}
