% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient.R
\name{nutrient}
\alias{nutrient}
\title{Get the sum of a single nutrient from a nutrition object}
\usage{
nutrient(obj, nutrient)
}
\arguments{
\item{obj}{The nutrition object -- a list returned by the `nutritionix_lookup()` function.}

\item{nutrient}{The nutrient to extract from the nutrition object. Must be one of: "calories", "fat", "saturated fat", "cholesterol", "sodium", "carbs", "fiber", "sugar", "potassium", etc... For a full list of supported nutrients, call `nutrients_list()`}
}
\description{
Gets the sum of a single nutrient from a nutrition object.
For example, gets the total calories or total fat.
}
\examples{

\dontrun{

obj <- nutritionix_lookup("1 cup of tofu")
nutrition(obj, "calories")

}

}
