% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutritionix_lookup.R
\name{nutritionix_lookup}
\alias{nutritionix_lookup}
\title{Get nutrition data using Nutritionix API}
\usage{
nutritionix_lookup(query)
}
\arguments{
\item{query}{The query to lookup nutrition information for}
}
\description{
Gets nutrition data from the Nutritionix API
}
\examples{

\dontrun{

nutritionix_lookup("1 cup of tofu")

}

}
