% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient.R
\name{nutrient}
\alias{nutrient}
\title{Get the sum of a single nutrient from a nutrition object}
\usage{
nutrient(obj, nutrient)
}
\arguments{
\item{obj}{The nutrition object -- a list returned by the `nutritionix_lookup()` function.}

\item{nutrient}{The nutrient to extract from the nutrition object. Must be one of the nutrients in `data(nutrients)`, such as : "calories", "fat", "carbs", "fiber", "sugar", etc}
}
\description{
Gets the sum of a single nutrient from a nutrition object. For example, gets the total calories or total fat. For a list of nutrients available, see `data(nutrients)`.
}
\examples{

\dontrun{

obj <- nutritionix_lookup("1 cheeseburger")
nutrition(obj, "calories")

}

}
